moroz_lesson_5_2 moved to other directory
